Logistic Regression & Tree-based Machine Learning
HR Department at Salifort Motors
Goal: Construct a prediction model that accurately predicts whether an employee will leave the company or not

Logistic Regression
The logistic regression model achieved precision of 80%, recall of 83%, f1-score of 80% (all weighted averages), and accuracy of 83%, on the test set.

Tree-based Machine Learning
After conducting feature engineering, the decision tree model achieved AUC of 93.8%, precision of 87.0%, recall of 90.4%, f1-score of 88.7%, and accuracy of 96.2%, on the test set. The random forest modestly outperformed the decision tree model.

The models and the feature importances extracted from the models confirm that employees at the company are overworked. To retain employees, recommendations could be presented to the stakeholders. Additionally, for another project, you could build a K-means model on this data and analyze the clusters to discover additional insights.